refactor(checker): split namespace display helper tests
