+++ crates/tsz-checker/src/state/type_analysis/mod.rs
@@ -8,6 +8,7 @@ mod computed_commonjs;
 pub(crate) mod computed_helpers;
 mod computed_helpers_binding;
 mod computed_helpers_circular;
+mod computed_helpers_namespace_display;
 mod computed_helpers_private;
 mod computed_loops;
 mod core;
